Convert kuksa integration into typescript
