+++ package.json
@@ -28,9 +28,9 @@
     "test-e2e-cypress": "cross-env TS_NODE_PROJECT= cypress run --project ./test/e2e --headless",
     "test-e2e": "cross-env TS_NODE_PROJECT=src/server/tsconfig.json NODE_ENV=test PORT=3005 ENABLE_OFFLINE_LOGIN=true start-test 'mkdir -p logs && npm start > ./logs/server-test.log 2>&1' 3005 test-e2e-cypress",
     "seed-database": "ts-node --project src/server/tsconfig.json scripts/seed-database.js",
-    "clear-temporary-tables": "ts-node --project src/server/tsconfig.json src/kuksa-integration/clear-temporary-tables.js",
-    "fetch-data": "ts-node --project src/server/tsconfig.json src/kuksa-integration/fetch-data.js",
-    "rebuild-tables": "ts-node --project src/server/tsconfig.json src/kuksa-integration/rebuild-tables.js",
+    "clear-temporary-tables": "ts-node --project src/server/tsconfig.json src/kuksa-integration/clear-temporary-tables.ts",
+    "fetch-data": "ts-node --project src/server/tsconfig.json src/kuksa-integration/fetch-data.ts",
+    "rebuild-tables": "ts-node --project src/server/tsconfig.json src/kuksa-integration/rebuild-tables.ts",
     "fetch-from-kuksa": "npm run clear-temporary-tables && npm run fetch-data && npm run rebuild-tables",
     "fetch-from-mock-kuksa": "cross-env KUKSA_API_PROXY_URL=\"\" KUKSA_API_USERNAME=username KUKSA_API_PASSWORD=password KUKSA_API_EVENTID=c3ef7c48-1df4-4bf8-b1d0-d523d601a3e1 KUKSA_API_ENDPOINT=http://localhost:4001 npm run fetch-from-kuksa",
     "mock-kuksa": "ts-node --project src/server/tsconfig.json test/utils/kuksa-integration/mock/mock-kuksa.js",
